medgemma-backend/requirements.txt

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.27.2
supabase>=2.27.3
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-dateutil==2.9.0
Pillow>=10.0.0
pydicom>=2.4.0
numpy>=1.24.0
pytest==8.3.4
pytest-asyncio==0.24.0